cloud and backend/remote: Broad mutexes to avoid a data race
The svchost package's disco.Disco and authentication cache components are
not concurrency-safe in the version of that library we use in Terraform
v1.5, and we can't upgrade without bringing in other dependency upgrades
that are not appropriate for the v1.5 series.

In practice the only situation where we can potentially make concurrent
calls to these is in the "remote" backend and the Cloud integration,
because they can potentially be involved in multiple concurrent
terraform_remote_state reads and/or state snapshot persist calls.
Therefore as a stopgap to reduce the potential for the race without
upgrading we'll mutex all of our uses of the service discovery object.

This is a far more coarse mutex than the one implemented in svchost itself
for Terraform v1.6 or later, but data races here can potentially entirely
crash Terraform and so it's warranted to be more cautious here even though
the race is relatively hard to hit in practice.

This temporary implementation has a notable limitation that it cannot
coordinate across multiple instances of the backend, and is therefore
useless and we need to find a different approach. Don't merge this.