rpcapi: Opening and closing stack configurations
Here we introduce the "Stacks" service to the terraform1 API, initially
focused on just the opening and closing and static analysis calls.

The implementation itself only supports Open/Close as of this commit,
which also introduced the idea of one handle depending on another so that
the latter can't be closed until the first is closed. This does now make
some of the handle-related operations fallible, so there are some API
changes to the handleTable API to deal with that.