rpcapi: codegen for RPC functions with streaming responses
Our generator for the wrapper stubs for the dynamically-instantiated gRPC
services could not previously handle the slightly different signature used
for the methods generated for RPC functions that return streaming
responses: they take a stream object as an argument instead of returning
a message directly.

Now we can generate wrapper stubs for those methods too, so that
subsequent commits can include the real implementations of those functions.
This commit includes just an unimplemented husk of Stacks.PlanStackChanges
as a placeholder to demonstrate that it's working.