rpcapi: Stacks.PlanStackChanges stub implementation
Our initial goal here is to blaze the trail through all of these different
API layers so that we can then fill out the details of each subsystem as
a concurrent development effort while relying on the natural interface
boundaries inherent in the system.

With that in mind, this introduces the most important parts of the
Stacks.PlanStackChanges API surface but with an implementation that behaves
as if no plan ever requires any changes. It just returns a warning
diagnostic so that once we have a real implementation later we'll be able
to easily recognize if we're accidentally talking to the stub by mistake.

The real implementation will follow in later commits.