rpcapi: Emit top-level tracing spans for requests
rpcapi is a normal gRPC server once it's running, so if the overall
Terraform CLI process has been configured to export telemetry then this
will emit a span for each incoming gRPC request.

In later commits we'll hopefully add further telemetry downstream to cover
interesting child spans such as running Terraform Core operations, but
this is at least enough to observe which requests a client is making
and what their outcomes were.