promising: Emit tracing spans for tasks and promises
Tasks and promises don't fit into the typical tree structure that
OpenTracing spans normally form, and so this follows the documented
conventions for coalesced batch processing making each task and promise
appear as a separate span but with links between the spans and events to
show when and how the tasks interacted with the promises.

Each promise is a span just because it acts as a convenient container to
attach an event log to. It doesn't actually do any work itself and so
a promise span will never have any children.

Tasks as spans is a more conventional use, allowing us to watch all of
the activities a task does in preparation for resolving each of the
promises that it's responsible for.

We use the span IDs as de-facto ids for the tasks and promises themselves,
since they have the appropriate amount of uniqueness for our purposes and
this saves us from devising our own separate ID scheme.