stackruntime: Helper for viewing traces in tests
Since our stackeval package is instrumented for OpenTelemetry tracing,
it's helpful to be able to see the trace events as part of the test log
when debugging.

tracesToTestLog arranges for any spans to be announced into the main test
log for the remainder of the current test, and then restores the original
trace provider afterwards as a test cleanup action.