promising: Once.Do must be called in a task context
Previously we tried to make this optional by wrapping the behavior of this
function in a call to promising.MainTask, but that's incorrect because
it causes the calling task to implicitly depend on the created main task
without any promise connecting them, which leads to a gap in the
promise-task bipartite graph. (Tasks can only depend on each other
indirectly through promises.)

It turns out that we don't actually currently even need this convenience
because all of the callers to promising.Once are in task context anyway.
Therefore we'll just remove this unnecessary extra wrapper and have the
calling task directly wait for the completion of the promise, thereby
causing the promise-task graph to be complete so we can avoid Once-related
deadlocks.