promising: Immediately fail all promises in a self-dependency chain
Previously we only directly signalled an error to whichever task was
unlucky enough to be the one completing the dependency cycle in the
promise-task graph.

Although that's technically correct if our only goal is to prevent the
formation of a blocking chain, in practice callers have their own concerns
that can be left unresolved if we rely on the self-dependency-related
errors to propagate organically as the callers each unwind themselves.

In particular, our "stackeval" package design relies on the idea of each
object only directly reporting its own errors and not reporting downstream
errors from dependent objects (since they should report themselves), but
that assumption is not valid if only one callstack directly sees the
ErrSelfDependent error, since the arbitrarily-selected observer might be
an indirect caller which discards errors on the assumption that another
observer will be the one to return them.

Instead then, here we introduce the idea of a promise having an "forced
error", which is generated by the promise system itself rather than by
the responsible task, and put all of the promises in a self-dependency
chain into that state as soon as the self-dependency is detected. This
causes the error to propagate up through _all_ of the calling stacks,
which means that at least one of them is likely to reach the top
caller but it also means that this particular class of error is likely
to be duplicated by multiple return paths.

Seeing multiple instances of the error is preferable to not seeing it at
all, so we'll accept that compromise and expect the caller to deduplicate
at the top join point if the multiple errors prove to be bothersome.

This also includes some additional tracing events that happened to be
useful while developing and debugging this change. For now the promising
package is intentionally very chatty in its trace output since it's useful
while the stackeval caller is under heavy development, but we will probably
tone this down some in later commits once things have settled more.