collections: Generic Set and Map types
This is a generalization of the similar types previously implemented in
package addrs, now available for a type in any package as long as we can
write a suitable rule for calculating a unique key for each distinct value
of that type.

Hopefully these eventually get replaced by future Go standard library
functions, but this is here to fill that gap for us in the meantime.