rpcapi: Resync the protobuf stubs for terraform1
Since we've been working in some long-lived branches outside of main
things have got a little tangled with changes to go.mod under rebasing,
with the generated terraform1 currently in our branch inconsistent with
the protoc-gen-go version that our go.mod currently selects.

This is just a boring change to reconcile that. It changes nothing
except a comment in the generated code.