stackruntime: Stub of "Plan" function
This establishes the overall API shape but implementation-wise just moves
the warning that the implementation is fake one level deeper, with
stackruntime now being the one to return that warning instead of the
rpcapi package directly.

Package rpcapi now monitors the two output channels from stackruntime.Plan
and translates what it sees into protobuf events to transmit to the
client. Of course given the above that's currently always just exactly
one diagnostic and no planned changes, but we'll start to make the
stackruntime.Plan function real in subsequent commits.