stackruntime: Allow hooking the start and end of the overall plan operation
This broad pair of hooks allows the caller (i.e. rpcapi) to generate a
parent tracing span around the entire plan operation, which all other
traced operations will descend from.

This is also in part just a way to show that the hook API is workable in
practice, since these are the first actual hooks.