stackplan: Expose the applyable flag as external description
Previously we retained the applyable flag only in the raw plan sequence,
but callers of PlanStackChange also need to know whether the plan they've
generated is applyable so that they can e.g. skip showing the plan
approval UI.

To support that we'll also expose the applyable flag as an
externally-consumable PlannedChange.Description variant, just encoding
the boolean result directly since there is no other metadata to
represent in this case.