stackeval: walkState with custom diagnostics handling
In an earlier commit we added walkState as something we could concisely
pass between all of the functions participating in a config/state walk
such as what we do in main_validating.go for validation, dealing with
the two cross-cutting concerns of firstly detecting when all of the
asynchronous operations have completed and secondly aggregating together
diagnostics produced by the various objects into a single set.

The diagnostics part of this problem was previously solved by just always
appending diagnostics into a hidden syncDiagnostics object. That works
okay for validation where we just wait until all the work is complete
and return all of the diagnostics together as a normal return value, but
for planning we expect to be able to stream out both planned changes and
diagnostics gradually as the (potentially-slow) planning process proceeds,
so the UI can give gradual feedback and thus feel more responsive than
just showing an indeterminite progress indicator for the entire operation.

To preserve the idea of just passing a single object through the walk
functions while giving a little more flexibility, here we introduce an
extra hooking capability where the plan phase can treat diagnostics
differently while still preserving the same external walkState API.
In future commits the plan walk driver functions will all share the same
walkState object and the entry point will populate it with the custom
callbacks required to create a bridge to the channel-based API for
emitting streaming diagnostics.

The old simple API is still available for the validate walk to use.
Instead of having the internal syncDiagnostics object directly inside
the walkState object, it now lives in the local scope created by calling
newWalkState, captured into the closures of the two callback functions.
The effect is equivalent.