stackruntime: Initial stubbing of Plan
This is just enough implementation of "Plan" for embedded stacks with
their own input variables and output values, and exposing the final
top-level output values as "planned changes".

This still doesn't really do anything end-user-useful but it's helpful to
have the expression evaluation and plan walk driver machinery in place
before moving on to the main useful language element: component blocks.
Support for components will follow in subsequent commits.