stacks/stackconfig: Hoist "typeexpr" out of the "internal" directory
Previously we placed "typeexpr" under "internal" to represent that it
wasn't for use by other parts of Terraform, but that was too strict: the
Stacks runtime needs to be able to import this package to access the
Defaults API.

For now we'll make it importable by the whole of Terraform. However, in
a later commit we might wish to establish an internal subdirectory directly
under "stacks" for packages that are only for stacks-related functionality,
since this "typeexpr" contains Stacks-specific logic and the rest of
Terraform ought to be using HCL's own "typeexpr" package instead.