stackeval: Very very basic component planning
This is the absolute bare-minimum integration with the main Terraform
language runtime to plan changes for a component. It's intentionally not
yet very robust and limited only to the built-in Terraform provider
because we're just trying to get the bones in place for seeing stack
planning working, rather than trying to build something complete enough
to ship.

This currently only deals with the resource instance changes that get
planned and doesn't expose the root module's output values for use
elsewhere in the stack configuration. Output value handling will follow
in subsequent commits.