staticcheck: Broadly ignore unused stuff in all of "stacks"
We're currently actively developing stuff under the "internal/stacks"
directory such that there are some loose ends not yet completed in order
to make our PRs at least _slightly_ more digestible.

For that reason we'll broadly turn off the "unused symbol" checks from
staticcheck under that entire directory for now, instead of making many
single-file exceptions that will be harder to clean up later. Once the
work in this directory settles we should remove this exception so that
the stacks-related packages are subject to the same rules as everything
else.

Unfortunately staticcheck can only work with configuration files if there
are no explicit checks on the command line, so this also makes the
top-level checks be a configuration file rather than just a command line
option in our staticcheck.sh script. The effect for the rest of the
codebase is the same, but having the top-level rules in a configuration
file means that subdirectories (like "internal/stacks") can inherit from
it to minimize the divergence from the typical rules.