Allow to provide multiple 'ADDR ID' pairs to be passed to `terraform import`
This seems to be needed, as at least in my work today I was unable to run `terraform import`
in succession and get it to actually import resources into a Terraform Cloud state: It would
always start out from the last refreshed/applied state.

In any case though: This looks like a trivial enough extension.