run copyright header check recursively
We have a few directories in this repo that will be remaining
MPL licensed (the provider protocol definitions). This is something
that the copywrite tool does not support out of the box - we need
to run the tool for each directory that contains a .copywrite.hcl
file in order to make sure that new files have the correct header.

To avoid adding yet more boilerplate with a `copyright_headers.go`
file in each package, let's move the copyright check to its own
make target and script the process of discovering all directories
with copywrite config.