rpcapi: Dependencies.CreateDependencyLocks
This new operation allows opening a dependency locks handle with lock
information provided inside the request, rather than creating one
indirectly from a dependency lock file.

This is intended to allow a caller to start with a dependency lock file,
saved the locked providers somewhere, and then in a later step using a
different Terraform Core RPC server instantiate another locks object with
the same providers without requiring access to the original dependency
lock file.

In future we might add a function for writing locks to disk as a dependency
lock file, which would then allow callers to construct entirely new
dependency lock files by providing the intended content themselves. We
don't have any direct need for that now, but that would complete this
family of functions if we added it later.