Build with Go 1.21.0 (#33691)
* Build with Go 1.21.0

* tests: update tls error message string

Go 1.21 introduced more detailed error conditions for mTLS connections.

As of Go 1.21, if the server is configured to require client auth, but no client cert is provided, the alert is now "certificate required", rather than "bad certificate". See https://github.com/golang/go/commit/62a994837a57a7d0c58bb364b580a389488446c9.

* declare go 1.21.0

As of Go 1.21.0, the go directive in go.mod declares a strict minimum
version of Go necessary to use a module. The go command will download a
copy of the newer toolchain if necessary.

* changelog: Go 1.21 supported OS versions

* scripts: pin staticcheck package

The last released version of staticcheck does not yet support Go
1.21.0, so pin to the latest commit.

When a version is released that supports Go 1.21.0, we should pin to
that instead.

* Go 1.21 contrib updates

* bump versions of hcl/v2 and go-cty

This update means that HCL, cty, and Terraform all use Unicode 15.

---------

Co-authored-by: Katy Moe <katy@katy.moe>
Co-authored-by: kmoe <5575356+kmoe@users.noreply.github.com>