cloud: Verify that state has been finalized
It's possible that state can be successfully uploaded to Terraform Cloud, but that the platform has an internal outage that prevents it from being acknowledged by the API.

This change verifies each state save arrives at the "finalized" state within 5 minutes. If that timeout is exceeded, an error is returned with detailed recovery instructions.