core: Allow passing in externally-configured provider instances
Normally Terraform Core expects what it sees as the root module to be
a self-contained root module which defines all of the provider
configurations it needs.

This commit is the start of making it valid for the apparent root module
to actually be a shared module, which means it'd expect to receive
provider configurations from its caller rather than defining them itself.

The caller of the "root" module is the Terraform Core caller written in
Go, so this is essentially a Go equivalent of the "providers" argument in
a "module" block in the Terraform language.