rpcapi: Provider plugin and root input value support
The Stacks.PlanStackChanges RPC function now accepts a dependency locks
handle, a provider cache handle, and a set of values for the root stack's
input variables.

All of the providers selected in the dependency locks are available for
use in the stack configuration as long as there is a suitable package for
all of them in the provider cache directory. The input values get
deserialized and then passed directly into stackruntime for use during
evaluation.