rpcapi: Won't be TFC-only forever
This is just a note to future maintainers to clarify the meaning of the
initial end-user-facing error message that says this interface is only
for Terraform Cloud.

Targeting Terraform Cloud as a client first is just a means to get
practical experience with this new style of integration before committing
to support it as a general API. We would still like to make this a
general-purpose integration point with compatibility guarantees in a later
release, once this approach is more proven.