stacks/stackruntime: Don't panic if component modules are invalid
ComponentConfig.RequiredProviderInstances was not considering the
possibility that ComponentConfig.ModuleTree will return nil if the
component's module tree is too invalid to load anything.

As usual in stackeval, we need to react to these situations by returning
safely to allow another goroutine to return the diagnostic messages that
caused the nil result.