stacks/stackruntime: Report if provider schema fetching fails
We were previously just ignoring the errors on the assumption that some
other codepath would propagate them, as is our assumption for all
inter-dependencies between objects in stackeval.

However, provider types are a bit of an oddity right now in that they are
really _only_ accessed by implication, and so we don't directly visit them
to get an opportunity to gather any errors. We'll hopefully improve on that
later, but for now we'll just emit the error once for each component
instance to ensure that it gets emitted _somewhere_.

We must also stop before trying to call terraform.NewContext if schema
loading failed, because otherwise terraform.NewContext itself will fail
trying to interact with a provider whose schema it wasn't given.