stackplan: Include schema correlation info in resource change descriptions
In order to find the appropriate schema to use when decoding the dynamic
values for a particular resource instance, callers will need to know which
provider's schema to consult, and which resource type and mode to consult
within that provider's schema.

We don't want callers to try to implement their own parsers for
Terraform's resource address syntax, and the provider information isn't
available any other way anyway, so we'll expose the three required lookup
keys as new fields in the events that report that a resource instance
change has been planned.