Add GraphNodeImportReferencer
When an import block is being evaluated by a resource within a module,
the references for the import evaluation must be scoped to the root
module where the block was written. In order to accomplish that without
adding a new import node type to the graph, we instead can opt to add a
new method which separates the import references from the config
references. The ReferenceTransformer can then use the correct root
module path to lookup the referenced subjects when they come from
ImportReferences as opposed to References.