rpcapi: NewInternalClient returns a socketless RPC client
Right now we are only using rpcapi from external callers inside Terraform
Cloud, but in the long run we will also want to expose equivalent
functionality through Terraform CLI.

rpcapi.NewInternalClient provides a relatively-lightweight way to achieve
that without requiring a child process and sockets. Instead, it just
allocates a buffer and passes gRPC messages between client and server
using that buffer.

This is of course still not as efficient as a direct function call, but
the kinds of operations that RPC API exposes tend to have far more
expensive elements than the RPC call itself, and using the RPC protocol
as the internal API should begin to establish a more explicit architecture
boundary between Terraform CLI and Terraform Core, so that clients of
the RPC API have the same level of access to Terraform Core as the CLI
workflow does.

So far we don't have any internal callers for this because nothing in the
RPC API has a corresponding CLI entry point. The first callers of this
internal RPC interface will probably be the commands providing the
CLI-driven development workflow for stacks.