rpcapi: Generalize the concurrency-safe gRPC streaming sender
Previously we had a wrapper specific to
terraform1.Stacks_PlanStackChangesServer, which was sufficient as long as
we had only implemented Stacks.PlanStackChanges.

In future commits we're going to implement Stacks.ApplyStackChanges too,
and so we'll need a similar wrapper for
terraform1.Stacks_ApplyStackChangesServer. The two have identical
implementation and differ only in the wrapped server interface, so
we can use a generic implementation and then alias it for convenient use
for the two currently-known usecases (to avoid writing out the horrendous
parameterized type expression more than we have to).