rpcapi: AppliedChange.ChangeDescription reserved field number
Correct behavior of the overall system relies on callers of
Stacks.ApplyStackChanges being able to faithfully preserve change
descriptions between runs even if they contain description fields the
client isn't yet familiar with.

To allow clients to explicitly test that capability, we'll reserve the
field number 20000 to guarantee it will never be used for a meaningful
field, so that client tests can produce a message with an encoded value
for that field and verify that it isn't lost during a round-trip.