stackeval: ChangeExec for orchestrating apply-time changes
Unlike the plan phase where we just visit everything in no particular
order and just let things be scheduled by the Go runtime wherever they
happen to get scheduled, we'd like a little more control over the parts
of the apply phase that might affect external infrastructure. In
particular, we'd like to see some top-level code that is clearly making
sure that all of the component instances from the plan are getting handled,
rather than leaving that implicit as part of the overall evaluation
process.

This new ChangeExec function (and associated supporting types) model an
explicit setup phase where we register all of the change actions we intend
to take, and then an execution phase where those actions can run
concurrently with other stackeval work while mediating through the
stackeval.Main type as usual.

This does require a small amount of tricky finagling to properly sequence
the steps so that we can freeze the set of change tasks before we start
executing but still give all of the tasks access to the appropriate
stackeval.Main object. As usual with this package, the idea is to keep
the weirdo promise-based scheduling stuff segregated from the main code
so that for everyday maintenence we can hopefully just focus on the
straight-through code that represents the main execution rules and worry
less about implementation details such as promises and tasks.