stackeval: ChangeExec returns states.State from component instance change
Previously we were returning just a value representing all of the
component's output values, but that's the wrong level of detail.

In the plan phase we have the corresponding function of ComponentInstance
return a plans.Plan object and then derive the object value representing
outputs from it only on request. Having the apply result be a *states.State
is the analogous level of detail, and then again we'll construct the
cty object representing all of the output values only on request during
expression evaluation.