stackeval: Factor out the two "walk drivers"
With validate, plan, and apply now all at least stubbed out we can see
that they all share some similar logic for visiting everything that's
relevant to an evaluation phase and gathering up diagnostics and other
external reports about each object.

Since it's important that the phases all visit the same objects so that
we can produce consistent results between phases, we'll accept a little
inversion-of-control complexity here in return for now having only two
"walk-driver" implementations: one for visiting the "static" objects and
one for visiting the "dynamic" objects.

 - The validate phase only visits static objects.
 - The plan phase visits static objects first, and then dynamic objects
   only if the static walk doesn't produce any errors.
 - The apply phase only visits the dynamic objects.

This also required some retroactive changes to some of my earlier work on
the validation phase since we can no longer assume that the validation
walk is the only one which visits the "static" objects. Their methods
now take EvalPhase arguments similar to those for the dynamic objects,
and we differentiate between the results in each phase so that we can
potentially add slight differences between the phases in future if needed.
As of this commit, though, the validate phase and the static portion of
the plan phase should produce identical results.